Bugfix (task descriptions, newlines)
Fix bug where change of a task's description would add a newline,
without a way to revert it due to how comparison of tasks works.

I am not happy with this fix, but it gets the job done. Maybe it's even
a feature that we are sloppy about trailing whitespace.